+++ R/argumentChecks.R
@@ -64,8 +64,8 @@ stab_conds_satisfied <- function(p, M, d, params, all_boldA=NULL, tolerance=1e-3
 #'    \item Lanne M., Virolainen S. 2025. A Gaussian smooth transition vector autoregressive model:
 #'       An application to the macroeconomic effects of severe weather shocks.
 #'       \emph{Journal of Economic Dynamics and Control}, \strong{178}, 105162.
-#'    \item Virolainen S. in press. Identification by non-Gaussianity in structural
-#'       smooth transition vector autoregressive models. \emph{Econometric Reviews}.
+#'    \item Virolainen S. 2026. Identification by non-Gaussianity in structural
+#'       smooth transition vector autoregressive models. \emph{Econometric Reviews}, \strong{45}:8, 1230-1248.
 #'  }
 #'  @keywords internal