+++ NEWS.md
@@ -125,7 +125,7 @@
 # sstvarss 1.2.2
 
 * The returned object from reorder_B_columns, swap_B_signs, and swap_parametrization now includes estimation results from all estimation rounds identically to
-  the original model (i.e., these results do now have the impact matrices nor the parametrization changed).
+  the original model (i.e., these results do now not have the impact matrices nor the parametrization changed).
 * Updated the vignette to match the current version of Virolainen (2025). 
 * Updated the data file usacpu to match the current version of Virolainen (2025).
 * Updated the reference to Lanne and Virolainen (2025).